+++ src-tauri/src/features/chat/model_runtime/provider_and_stream/types.rs
@@ -21,7 +21,6 @@ const SCREENSHOT_ARTIFACT_MAX_ITEMS: usize = 24;
 
 type ReadFileMcpClient = rmcp::service::RunningService<rmcp::RoleClient, ()>;
 type OperateMcpClient = rmcp::service::RunningService<rmcp::RoleClient, ()>;
-type TodoMcpClient = rmcp::service::RunningService<rmcp::RoleClient, ()>;
 
 struct RuntimeToolAssembly {
     tools: Vec<Box<dyn RuntimeToolDyn>>,
@@ -29,5 +28,4 @@ struct RuntimeToolAssembly {
     unavailable_tool_notices: Vec<String>,
     _mcp_read_file_client: Option<ReadFileMcpClient>,
     _mcp_operate_client: Option<OperateMcpClient>,
-    _mcp_todo_client: Option<TodoMcpClient>,
 }