+++ src-tauri/src/features/system/tools/terminal/exec.rs
@@ -497,7 +497,7 @@ mod terminal_exec_tests {
             dequeue_lock: Arc::new(Mutex::new(())),
             delegate_runtime_threads: Arc::new(Mutex::new(HashMap::new())),
             delegate_recent_threads: Arc::new(Mutex::new(VecDeque::new())),
-            provider_streaming_disabled_keys: Arc::new(Mutex::new(HashSet::new())),
+            provider_streaming_disabled_keys: Arc::new(Mutex::new(HashMap::new())),
             provider_system_message_user_fallback_keys: Arc::new(Mutex::new(HashSet::new())),
             hidden_skill_snapshot_cache: Arc::new(Mutex::new(String::new())),
             preferred_release_source: Arc::new(Mutex::new("github".to_string())),